Fully upgraded assembly machine to i64.

+++ helpers/rust/src/machine/hook.rs
@@ -13,7 +13,7 @@ pub enum HookResult {
     Skip,
 
     /// Go to a specific instruction.
-    Goto(i32),
+    Goto(i64),
 
     /// Abort the entire program.
     Abort,
